+++ kubernetes-client-api/src/test/java/io/fabric8/kubernetes/client/internal/UtilsTest.java
@@ -76,8 +76,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.SortedMap;
-import java.util.TreeMap;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
@@ -173,65 +171,6 @@ class UtilsTest {
     final String result = Utils.interpolateString(input, parameters);
     // Then
     assertEquals("This is a \"template string\" and the following is code ${NOT_REPLACED}: '1' === '1'; /* END */", result);
-  }
-
-  @Test
-  void testGetPluralFromKind() {
-    // Given
-    SortedMap<String, Class> pluralToKubernetesResourceMap = new TreeMap<>();
-    pluralToKubernetesResourceMap.put("bindings", Binding.class);
-    pluralToKubernetesResourceMap.put("componentstatuses", ComponentStatus.class);
-    pluralToKubernetesResourceMap.put("configmaps", ConfigMap.class);
-    pluralToKubernetesResourceMap.put("endpoints", Endpoints.class);
-    pluralToKubernetesResourceMap.put("events", Event.class);
-    pluralToKubernetesResourceMap.put("limitranges", LimitRange.class);
-    pluralToKubernetesResourceMap.put("namespaces", Namespace.class);
-    pluralToKubernetesResourceMap.put("nodes", Node.class);
-    pluralToKubernetesResourceMap.put("persistentvolumeclaims", PersistentVolumeClaim.class);
-    pluralToKubernetesResourceMap.put("persistentvolumes", PersistentVolume.class);
-    pluralToKubernetesResourceMap.put("pods", Pod.class);
-    pluralToKubernetesResourceMap.put("podtemplates", PodTemplate.class);
-    pluralToKubernetesResourceMap.put("replicationcontrollers", ReplicationController.class);
-    pluralToKubernetesResourceMap.put("resourcequotas", ResourceQuota.class);
-    pluralToKubernetesResourceMap.put("secrets", Secret.class);
-    pluralToKubernetesResourceMap.put("serviceaccounts", ServiceAccount.class);
-    pluralToKubernetesResourceMap.put("services", Service.class);
-    pluralToKubernetesResourceMap.put("mutatingwebhookconfigurations", MutatingWebhookConfiguration.class);
-    pluralToKubernetesResourceMap.put("validatingwebhookconfigurations", ValidatingWebhookConfiguration.class);
-    pluralToKubernetesResourceMap.put("customresourcedefinitions", CustomResourceDefinition.class);
-    pluralToKubernetesResourceMap.put("controllerrevisions", ControllerRevision.class);
-    pluralToKubernetesResourceMap.put("daemonsets", DaemonSet.class);
-    pluralToKubernetesResourceMap.put("deployments", Deployment.class);
-    pluralToKubernetesResourceMap.put("replicasets", ReplicaSet.class);
-    pluralToKubernetesResourceMap.put("statefulsets", StatefulSet.class);
-    pluralToKubernetesResourceMap.put("tokenreviews", TokenReview.class);
-    pluralToKubernetesResourceMap.put("localsubjectaccessreviews", LocalSubjectAccessReview.class);
-    pluralToKubernetesResourceMap.put("selfsubjectaccessreviews", SelfSubjectAccessReview.class);
-    pluralToKubernetesResourceMap.put("selfsubjectrulesreviews", SelfSubjectRulesReview.class);
-    pluralToKubernetesResourceMap.put("subjectaccessreviews", SubjectAccessReview.class);
-    pluralToKubernetesResourceMap.put("horizontalpodautoscalers", HorizontalPodAutoscaler.class);
-    pluralToKubernetesResourceMap.put("cronjobs", CronJob.class);
-    pluralToKubernetesResourceMap.put("jobs", Job.class);
-    pluralToKubernetesResourceMap.put("certificatesigningrequests", CertificateSigningRequest.class);
-    pluralToKubernetesResourceMap.put("leases", Lease.class);
-    pluralToKubernetesResourceMap.put("endpointslices", EndpointSlice.class);
-    pluralToKubernetesResourceMap.put("ingresses", Ingress.class);
-    pluralToKubernetesResourceMap.put("networkpolicies", NetworkPolicy.class);
-    pluralToKubernetesResourceMap.put("poddisruptionbudgets", PodDisruptionBudget.class);
-    pluralToKubernetesResourceMap.put("podsecuritypolicies", PodSecurityPolicy.class);
-    pluralToKubernetesResourceMap.put("clusterrolebindings", ClusterRoleBinding.class);
-    pluralToKubernetesResourceMap.put("clusterroles", ClusterRole.class);
-    pluralToKubernetesResourceMap.put("rolebindings", RoleBinding.class);
-    pluralToKubernetesResourceMap.put("roles", Role.class);
-    pluralToKubernetesResourceMap.put("priorityclasses", PriorityClass.class);
-    pluralToKubernetesResourceMap.put("csidrivers", CSIDriver.class);
-    pluralToKubernetesResourceMap.put("csinodes", CSINode.class);
-    pluralToKubernetesResourceMap.put("storageclasses", StorageClass.class);
-    pluralToKubernetesResourceMap.put("volumeattachments", VolumeAttachment.class);
-
-    // When & Then
-    pluralToKubernetesResourceMap.forEach(
-        (plural, kubernetesResource) -> assertEquals(plural, Utils.getPluralFromKind(kubernetesResource.getSimpleName())));
   }
 
   @Test